feat: dynamic node layout positioning algorithm
Remove the hardcoded 120 pixel node height and make it dynamic instead.

Add a method that gets called when a node gets a new port to determine
if two adjacent nodes in a column do not maintain a minimum gap. It will
recursively move nodes below downwards if there is not enough space.

Add constants.rs to manage constants.